+++ Sources/SentryObjCCompat/SentryObjCHub.swift
@@ -127,6 +127,10 @@ import Foundation
         wrapped.add(crumb.wrapped)
     }
 
+    @objc(addFeatureFlagWithName:result:) public func addFeatureFlag(name: String, result: Bool) {
+        wrapped.addFeatureFlag(name: name, result: result)
+    }
+
     @objc public func getClient() -> SentryObjCClient? {
         guard let client = wrapped.getClient() else { return nil }
         return SentryObjCClient(client)